Detect rustc version, set nightly with rustup otherwise (fixes #35)
